requirements.txt
flask
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pytesseract==0.3.10